% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hestia_functions.R
\name{get_transmission_details}
\alias{get_transmission_details}
\title{Create Transmission Probability Matrix}
\usage{
get_transmission_details(inf_model)
}
\arguments{
\item{inf_model}{infection process model object yielded by
make_infection_model()}
}
\value{
A list with the following entries:
\itemize{
\item "states", a character vector of state names
\item "trans_matrix", a matrix with transition rate values between origin
(column) and destination (row) compartments where given. Where no values
are provided for transition rates the entry is equal to 1e-10.
\item "mult_matrix", a matrix with split values for the transition between
origin (column) and destination (row) compartments where given. Where no
split values are provided for transition rates the entry is equal to 1.
\item "trans_to_fit", a data frame with information on all transitions within
the model, with parameter indexing details if the transition rate is being
fit
\item "mult_to_fit", a data frame with information on all splits specified in
the model, with parameter indexing details if the split is being fit
\item "inf_states", numeric values corresponding to the indices in "states"
which are the infectious states
\item "mult_inf_probs", indicator for whether infectious states have shared or
separate inta-household infection probabilities. If FALSE then all infection
probabilities are shared across infectious compartments.
}
}
\description{
Create Transmission Probability Matrix
}
