% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hestia-package.R
\docType{package}
\name{hestia-package}
\alias{hestia-package}
\alias{hestia}
\title{The 'hestia' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.7. https://mc-stan.org
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
