% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hestia_functions.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse Logit Transformation}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{Numeric value}
}
\value{
A numeric value between 0 and 1.
}
\description{
Inverse Logit Transformation
}
\keyword{internal}
