% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hestia_functions.R
\name{logit}
\alias{logit}
\title{Logit Transformation}
\usage{
logit(x)
}
\arguments{
\item{x}{Numeric value between 0 and 1}
}
\value{
A numeric value.
}
\description{
Logit Transformation
}
\keyword{internal}
