% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hestia_functions.R
\name{make_infection_model}
\alias{make_infection_model}
\title{Build Infection Process Model}
\usage{
make_infection_model(..., mult_inf_probs = FALSE)
}
\arguments{
\item{...}{a series of \link{progress} or \link{transmit} function calls}

\item{mult_inf_probs}{If FALSE then all infection probabilities are shared
across infectious compartments}
}
\value{
A data frame with a row for each unique transition in the infection
process model and the following columns:
\itemize{
\item "from", the name of the origin compartment
\item "to", the name of the destination compartment
\item "source", NA (so output columns match with \code{transmit})
\item "rate_name", name of transition rate parameter if fitting
\item "rate_value", numeric value of transition rate if not fitting
\item "split_name", name of split parameter if fitting
\item "split_value", numeric value of split if not fitting
\item "source", names of compartment(s) that are the source of new infections
\item "mult_inf_probs", a logical indicator for whether infectious compartments
have separate (\code{true}) or shared (\code{FALSE}) intra-household infection
probabilities
}
}
\description{
Build Infection Process Model
}
\examples{
# Basic SIR model with recovery rate gamma to be fit
make_infection_model(transmit(from = "S", to = "I"),
                     progress(from = "I", to = "R", gamma = NA))

}
