% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hestia_functions.R
\name{rename_chains}
\alias{rename_chains}
\title{Extract chains and renames with user-provided parameter names}
\usage{
rename_chains(inf_model, model_output)
}
\arguments{
\item{inf_model}{Infection process model object generated by
\link{make_infection_model}}

\item{model_output}{A list generated within \link{run_model} which contains
the stanfit, the stan input data, and covariate names}
}
\value{
A \code{draws_array} object with chains for each model parameter
}
\description{
Extract chains and renames with user-provided parameter names
}
\keyword{internal}
