% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hestia_functions.R
\name{run_model}
\alias{run_model}
\title{Run Household Transmission Model}
\usage{
run_model(
  inf_model,
  obs_model,
  data,
  init_probs,
  epsilon = 1e-10,
  ih_cov = NULL,
  eh_cov = NULL,
  iter = 2000,
  chains = 4,
  cores = getOption("mc.cores", 1L),
  init = NULL,
  save_states = FALSE
)
}
\arguments{
\item{inf_model}{Infection process model object generated by
\link{make_infection_model}}

\item{obs_model}{Observation process model object generated by
\link{make_observation_model}}

\item{data}{data frame with columns the following columns:
\itemize{
\item "t", time of observation
\item "hh_id", numeric household ID which must be 1, 2, ..., N where N is the
total number of household
\item "part_id", participant ID within the household which must be 1, 2, ... n
where n is the number of observed household members
\item "hh_size", number of household members (whether observed or not)
\item At least one column with binary outcome information where the column
name matches the outcome name provided to \link{make_observation_model}
}}

\item{init_probs}{vector of initial probabilities for each infection process
model state}

\item{epsilon}{very small number to use for zero probability transitions}

\item{ih_cov}{NULL for run without covariates, otherwise data frame with
intra-household covariates for each participant}

\item{eh_cov}{NULL for run without covariates, otherwise data frame with
extra-household covariates for each participant}

\item{iter}{number of MCMC iterations}

\item{chains}{number of MCMC chains}

\item{cores}{number of cores for parallelization}

\item{init}{initial conditions for MCMC chains}

\item{save_states}{indicator for whether to save state probabilities}

\item{file}{file path to stan model}
}
\value{
\code{draws_array} object with chains for each model parameter
}
\description{
Run Household Transmission Model
}
\examples{
# Subset sir package data to 10 households (for speedy example)
sir_sub <- sir[sir$hh_id <= 10, ]

# Make infetion process model for SIR model
inf_mod <- make_infection_model(
  transmit(from = "S", to = "I"),
  progress(from = "I", to = "R", gamma = NA))

# Make observation process model
obs_mod <- make_observation_model(
  pcr = c("S" = 0.05, "I" = 0.95, "R" = 0.05),
  igg = c("S" = 0.01, "I" = 0.01, "R" = 0.8))

run_model(inf_model = inf_mod,
          obs_model = obs_mod,
          data = sir_sub,
          init_probs = c(1 - 2 * 1e-10, 1e-10, 1e-10))

}
