% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hestia_functions.R
\name{split_check}
\alias{split_check}
\title{Checking split parameter specification}
\usage{
split_check(to, split)
}
\arguments{
\item{to}{Character vector giving the name(s) of the destination compartment}

\item{split}{Numeric or character vector indicating how people moving out of
the starting compartment are split between the destination compartments}
}
\description{
Utility function for checking whether split parameter is
properly defined
}
\keyword{internal}
