% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hestia_functions.R
\name{to_from_check}
\alias{to_from_check}
\title{Checking to and from parameter specification}
\usage{
to_from_check(to, from)
}
\arguments{
\item{to}{Character vector giving the name(s) of the destination compartment}

\item{from}{string giving name of origin compartment}
}
\description{
Utility function for checking whether to and from parameters
are properly defined
}
\keyword{internal}
