% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hestia_functions.R
\name{get_transmission_details}
\alias{get_transmission_details}
\title{Creates transmission probability matrix,}
\usage{
get_transmission_details(inf_model)
}
\arguments{
\item{inf_model}{infection process model object yielded by make_infection_model()}
}
\description{
Creates transmission probability matrix,
}
