% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hestia-package.R
\docType{package}
\name{hestia-package}
\alias{hestia}
\alias{hestia-package}
\title{hestia: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
