% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hestia_functions.R
\name{make_infection_model}
\alias{make_infection_model}
\title{Build Infection Process Model}
\usage{
make_infection_model(..., mult_inf_probs = FALSE)
}
\arguments{
\item{...}{a series of \link{progress} or \link{transmit} function calls}

\item{mult_inf_probs}{If FALSE then all infection probabilities are shared
across infectious compartments}
}
\description{
Build Infection Process Model
}
