% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hestia_functions.R
\name{make_infection_model}
\alias{make_infection_model}
\title{Builds infection process model.}
\usage{
make_infection_model(..., mult_inf_probs = FALSE)
}
\arguments{
\item{...}{a series of transit or transmit function calls}

\item{ih_cov}{indicator for whether intra-household infection risk is a function of covariates}

\item{eh_cov}{indicator for whether extra-household infection risk is a function of covariates}
}
\description{
Builds infection process model.
}
