% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hestia_functions.R
\name{make_observation_model}
\alias{make_observation_model}
\title{Make Observation Process Model}
\usage{
make_observation_model(...)
}
\arguments{
\item{...}{A series of named vectors. Each vector corresponds to an observation
type. Each entry in the vector is named for the corresponding compartment in
the infection process model. The value is the proability of observing a positive
observation given the individual is in the compartment.}
}
\description{
Make Observation Process Model
}
