% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hestia_functions.R
\name{make_stan_data}
\alias{make_stan_data}
\title{Create Data for Stan Model}
\usage{
make_stan_data(
  inf_model,
  obs_model,
  data,
  init_probs,
  epsilon = 1e-10,
  ih_cov = NULL,
  eh_cov = NULL
)
}
\arguments{
\item{inf_model}{Infection process model object generated by \link{make_infection_model}}

\item{obs_model}{Observation process model object generated by \link{make_observation_model}}

\item{data}{data frame with columns....}

\item{init_probs}{vector of initial probabilities for each infection process model state}

\item{epsilon}{very small number to use for zero probability transitions}

\item{ih_cov}{NULL for run without covariates, otherwise data frame with intra-household covariates for each participant}

\item{eh_cov}{NULL for run without covariates, otherwise data frame with extra-household covariates for each participant}
}
\description{
Create Data for Stan Model
}
