% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hestia_functions.R
\name{run_model}
\alias{run_model}
\title{Run Household Transmission Model}
\usage{
run_model(
  inf_model,
  obs_model,
  data,
  init_probs,
  epsilon = 1e-10,
  ih_cov = NULL,
  eh_cov = NULL,
  file = "stan/hmm.stan",
  iter = 2000,
  chains = 4,
  cores = getOption("mc.cores", 1L),
  init = NULL,
  save_states = FALSE
)
}
\arguments{
\item{inf_model}{Infection process model object generated by \link{make_infection_model}}

\item{obs_model}{Observation process model object generated by \link{make_observation_model}}

\item{data}{data frame with columns....}

\item{init_probs}{vector of initial probabilities for each infection process model state}

\item{epsilon}{very small number to use for zero probability transitions}

\item{ih_cov}{NULL for run without covariates, otherwise data frame with intra-household covariates for each participant}

\item{eh_cov}{NULL for run without covariates, otherwise data frame with extra-household covariates for each participant}

\item{file}{file path to stan model}

\item{iter}{number of MCMC iterations}

\item{chains}{number of MCMC chains}

\item{cores}{number of cores for parallelization}

\item{init}{initial conditions for MCMC chains}

\item{save_states}{indicator for whether to save state probabilities}
}
\description{
Run Household Transmission Model
}
