% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hestia_functions.R
\name{split_check}
\alias{split_check}
\title{Utility function for checking whether split parameter is properly defined}
\usage{
split_check(to, split)
}
\arguments{
\item{to}{character vector giving the name(s) of the destination compartment}

\item{split}{numeric or character vector indicating how people moving out of the starting compartment are split between the destination compartments}
}
\description{
Utility function for checking whether split parameter is properly defined
}
