% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hestia_functions.R
\name{transit}
\alias{transit}
\title{Defines a state transition in the infection process model which does not
represent an transmission (infection) event}
\usage{
transit(from, to, split = NA, ...)
}
\arguments{
\item{from}{string giving name of origin compartment}

\item{to}{string giving the name of the destination compartment}
}
\description{
Defines a state transition in the infection process model which does not
represent an transmission (infection) event
}
