% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hestia_functions.R
\name{transmit}
\alias{transmit}
\title{Defines a state transition in the infection process model which is
the result of a transmission (infection) event}
\usage{
transmit(from, to, source = NA, split = NA)
}
\arguments{
\item{from}{string giving name of origin compartment}

\item{to}{string giving the name of the destination compartment}

\item{source}{string (or vector of strings) designating which compartments are infectious. If NULL, the destination compartment is presumed to be the infectious compartment.}
}
\description{
Defines a state transition in the infection process model which is
the result of a transmission (infection) event
}
